requirements.txt
asn1crypto==1.5.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
docusign-esign==3.22.0
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
nose==1.3.7
numpy==1.24.3
opencv-python==4.7.0.72
oscrypto==1.3.0
pdf2image==1.16.3
Pillow==9.5.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
requests==2.30.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==2.0.2
Werkzeug==2.3.4
